requirements.txt
Flask>=2.3.3
Flask-Compress>=1.23
Flask-SocketIO>=5.6.0
Flask-SQLAlchemy>=3.1.1
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.1.0
google-generativeai>=0.8.0
yfinance>=1.0.0
feedparser>=6.0.0
paramiko>=3.0.0
edge-tts>=7.0.0
ytmusicapi>=1.3.0
requests>=2.31.0
psutil>=5.9.0
tzlocal>=5.0
pytz>=2024.1
